+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `Select` now supports grouped options with a `title` and option `description`, as well as `BoxProps` for layout styling.
+- New components: `InputDate`, `Calendar`
 
 ### Changed